0.1.7: 根据baseCtrler重构，尚未测试完成

+++ ctrlers/index.js
@@ -1,26 +1,47 @@
-var Ctrler = function(model) {
-    this.model = (model) ? model : null;
+var matcher = function(id) {
+    return (id && id.match(/^[0-9a-fA-F]{24}$/));
 }
+
+var Ctrler = function(model) {
+    this.model = model ? model : null;
+}
+
+Ctrler.prototype.checkId = matcher;
 
 Ctrler.prototype.create = function(baby, callback) {
-    var baby = new this.model(baby);
-    baby.save(function(err) {
-        callback(err, baby);
-    });
+    this.model.create(baby, callback);
 }
 
-Ctrler.prototype.read = function(id, callback) {
-    if (id && id.match(/^[0-9a-fA-F]{24}$/)) {
-        this.model.findById(id).exec(callback);
-    }
+Ctrler.prototype.findById = function(id, callback) {    
+    if (matcher(id)) this.model.findById(id).exec(callback);
 }
 
-Ctrler.prototype.update = function(id, baby, callback) {
-    this.model.findByIdAndUpdate(id, baby, callback);
+Ctrler.prototype.update = function(id, update, callback) {
+    if (matcher(id)) this.model.findByIdAndUpdate(id, update, callback);
+}
+
+Ctrler.prototype.updates = function(params, update, callback) {
+    if (matcher(id)) this.model.update(params, update, callback);
+}
+
+Ctrler.prototype.updateOne = function(id, update, callback) {
+    if (matcher(id)) this.model.findOneAndUpdate(id, update, callback);
 }
 
 Ctrler.prototype.remove = function(id, callback) {
-    this.model.findByIdAndRemove(id, callback);
+    if (matcher(id)) this.model.findByIdAndRemove(id, callback);
+}
+
+Ctrler.prototype.removes = function(params, callback) {
+    if (matcher(id)) this.model.remove(params, callback);
+}
+
+Ctrler.prototype.removeOne = function(id, callback) {
+    if (matcher(id)) this.model.findOneAndRemove(id, callback);
+}
+
+Ctrler.prototype.populate = function(doc, params, callback) {
+    this.model.populate(doc, params, callback);
 }
 
 Ctrler.prototype.count = function(params, callback) {
@@ -51,23 +72,6 @@ Ctrler.prototype.page = function(page, limit, params, callback) {
             callback(err);
         }
     });
-}
-
-// advanced find
-Ctrler.prototype.find = function(params, callback) {
-    this.model.find(params).exec(callback);
-}
-
-// advanced findOne
-Ctrler.prototype.findOne = function(params, callback) {
-    this.model.findOne(params).exec(callback);
-}
-
-// advanced query:
-// var demo = user.query('findOne',{name: 'demo'});
-// demo.select('sth'); demo.exec(function(err,result){});
-Ctrler.prototype.query = function(type, params) {
-    return this.model[type](params);
 }
 
 module.exports = Ctrler;
